+++ iroh-net/src/relay/http/server.rs
@@ -23,13 +23,11 @@ use tracing::{debug, debug_span, error, info, info_span, warn, Instrument};
 use tungstenite::handshake::derive_accept_key;
 
 use crate::key::SecretKey;
-use crate::relay::http::{
-    HTTP_UPGRADE_PROTOCOL, SUPPORTED_WEBSOCKET_VERSION, WEBSOCKET_UPGRADE_PROTOCOL,
-};
+use crate::relay::http::SUPPORTED_WEBSOCKET_VERSION;
 use crate::relay::server::{ClientConnHandler, MaybeTlsStream};
 use crate::relay::MaybeTlsStreamServer;
 
-use super::{LEGACY_RELAY_PATH, RELAY_PATH};
+use super::{Protocol, LEGACY_RELAY_PATH, RELAY_PATH};
 
 type BytesBody = http_body_util::Full<hyper::body::Bytes>;
 type HyperError = Box<dyn std::error::Error + Send + Sync>;
@@ -60,54 +58,21 @@ fn downcast_upgrade(upgraded: Upgraded) -> Result<(MaybeTlsStream, Bytes)> {
     }
 }
 
-/// The HTTP upgrade protocol used for relaying.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum Protocol {
-    /// Relays over the custom relaying protocol with a custom HTTP upgrade header.
-    Relay,
-    /// Relays over websockets.
-    ///
-    /// Originally introduced to support browser connections.
-    Websocket,
-}
+/// The server HTTP handler to do HTTP upgrades
+async fn relay_connection_handler(
+    protocol: Protocol,
+    conn_handler: &ClientConnHandler,
+    upgraded: Upgraded,
+) -> Result<()> {
+    debug!(?protocol, "relay_connection upgraded");
+    let (io, read_buf) = downcast_upgrade(upgraded)?;
+    ensure!(
+        read_buf.is_empty(),
+        "can not deal with buffered data yet: {:?}",
+        read_buf
+    );
 
-impl Protocol {
-    /// The HTTP upgrade header used or expected
-    pub const fn upgrade_header(&self) -> &'static str {
-        match self {
-            Protocol::Relay => HTTP_UPGRADE_PROTOCOL,
-            Protocol::Websocket => WEBSOCKET_UPGRADE_PROTOCOL,
-        }
-    }
-
-    /// Tries to match the value of an HTTP upgrade header to figure out which protocol should be initiated.
-    pub fn parse_header(header: &HeaderValue) -> Option<Self> {
-        let header_bytes = header.as_bytes();
-        if header_bytes == Protocol::Relay.upgrade_header().as_bytes() {
-            Some(Protocol::Relay)
-        } else if header_bytes == Protocol::Websocket.upgrade_header().as_bytes() {
-            Some(Protocol::Websocket)
-        } else {
-            None
-        }
-    }
-
-    /// The server HTTP handler to do HTTP upgrades
-    async fn relay_connection_handler(
-        self,
-        conn_handler: &ClientConnHandler,
-        upgraded: Upgraded,
-    ) -> Result<()> {
-        debug!(protocol = ?self, "relay_connection upgraded");
-        let (io, read_buf) = downcast_upgrade(upgraded)?;
-        ensure!(
-            read_buf.is_empty(),
-            "can not deal with buffered data yet: {:?}",
-            read_buf
-        );
-
-        conn_handler.accept(self, io).await
-    }
+    conn_handler.accept(protocol, io).await
 }
 
 /// The Relay HTTP server.
@@ -487,9 +452,12 @@ impl Service<Request<Incoming>> for ClientConnHandler {
                     async move {
                         match hyper::upgrade::on(&mut req).await {
                             Ok(upgraded) => {
-                                if let Err(e) = protocol
-                                    .relay_connection_handler(&closure_conn_handler, upgraded)
-                                    .await
+                                if let Err(e) = relay_connection_handler(
+                                    protocol,
+                                    &closure_conn_handler,
+                                    upgraded,
+                                )
+                                .await
                                 {
                                     warn!(
                                         "upgrade to \"{}\": io error: {:?}",
